requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
PyJWT>=2.8
psycopg2-binary>=2.9
pytest>=8.0
pytest-django>=4.8
